Download JavaScript earlier.
What
===
Move the JavaScript tag to the HEAD so that it enters the queue for
loading earlier, with the defer async attribute, so that it doesn't
execute until the DOM is prepared.

Why
===
The JavaScript tags current location means it doesn't start downloading
until the DOM is loaded, which delays execution. The code needs to
execute after DOM load though because it queries the DOM, and adding the
defer tag ensures execution doesn't happen until the DOM is loaded
making this change safe for modern browsers.